downloads can now be removed (staging and db)

+++ coastal-hazards-portal/src/main/java/gov/usgs/cida/coastalhazards/rest/data/DownloadResource.java
@@ -54,16 +54,16 @@ public class DownloadResource {
     @Produces("application/zip")
     public Response downloadItem(@PathParam("id") String id) throws IOException {
         Response response = null;
-        try (ItemManager itemManager = new ItemManager()) {
+        try (ItemManager itemManager = new ItemManager(); DownloadManager downloadManager = new DownloadManager()) {
             Item item = itemManager.load(id);
             if (item == null) {
                 throw new NotFoundException();
             } else {
                 File zipFile = null;
-                DownloadManager manager = new DownloadManager();
+                
                 try {
-                    if (manager.isPersisted(id)) {
-                        Download persistedDownload = manager.load(id);
+                    if (downloadManager.isPersisted(id)) {
+                        Download persistedDownload = downloadManager.load(id);
                         // if we switch this to external file server or S3,
                         // redirect to this uri as a url
                         zipFile = persistedDownload.fetchZipFile();
@@ -84,7 +84,7 @@ public class DownloadResource {
                         } catch (URISyntaxException ex2) {
                             throw new DownloadStagingUnsuccessfulException();
                         }
-                        manager.save(download);
+                        downloadManager.save(download);
                     } else {
                         throw new DownloadStagingUnsuccessfulException();
                     }
@@ -110,40 +110,42 @@ public class DownloadResource {
     public Response getSession(@PathParam("id") String id) throws IOException, NoSuchAlgorithmException {
         Response response = null;
         String sessionJSON = sessionManager.load(id);
-        if (sessionJSON == null) {
-            throw new NotFoundException();
-        } else {
-            DownloadManager manager = new DownloadManager();
-            File zipFile = null;
-            try {
-                if (manager.isPersisted(id)) {
-                    Download download = manager.load(id);
-                    zipFile = new File(new URI(download.getPersistanceURI()));
-                    if (zipFile == null || !zipFile.exists()) {
+        try ( DownloadManager downloadManager = new DownloadManager()) {
+            if (sessionJSON == null) {
+                throw new NotFoundException();
+            } else {
+
+                File zipFile = null;
+                try {
+                    if (downloadManager.isPersisted(id)) {
+                        Download download = downloadManager.load(id);
+                        zipFile = new File(new URI(download.getPersistanceURI()));
+                        if (zipFile == null || !zipFile.exists()) {
+                            throw new FileNotFoundException();
+                        }
+                    } else {
                         throw new FileNotFoundException();
                     }
-                } else {
-                    throw new FileNotFoundException();
-                }
-            } catch (FileNotFoundException| URISyntaxException ex) {
-                Session session = Session.fromJSON(sessionJSON);
-                File stagingDir = DownloadUtility.createDownloadStagingArea();
-                boolean staged = DownloadUtility.stageSessionDownload(session, stagingDir);
-                if (staged) {
-                    Download download = DownloadUtility.zipStagingAreaForDownload(stagingDir);
-                    download.setSessionId(id);
-                    try {
-                        zipFile = download.fetchZipFile();
-                    } catch (URISyntaxException ex2) {
+                } catch (FileNotFoundException| URISyntaxException ex) {
+                    Session session = Session.fromJSON(sessionJSON);
+                    File stagingDir = DownloadUtility.createDownloadStagingArea();
+                    boolean staged = DownloadUtility.stageSessionDownload(session, stagingDir);
+                    if (staged) {
+                        Download download = DownloadUtility.zipStagingAreaForDownload(stagingDir);
+                        download.setSessionId(id);
+                        try {
+                            zipFile = download.fetchZipFile();
+                        } catch (URISyntaxException ex2) {
+                            throw new DownloadStagingUnsuccessfulException();
+                        }
+                        downloadManager.save(download);
+                    } else {
                         throw new DownloadStagingUnsuccessfulException();
                     }
-                    manager.save(download);
-                } else {
-                    throw new DownloadStagingUnsuccessfulException();
                 }
+                String contentDisposition = "attachment; filename=\"" + id + ".zip\"";
+                response = Response.ok(zipFile, "application/zip").header("Content-Disposition",  contentDisposition).build();
             }
-            String contentDisposition = "attachment; filename=\"" + id + ".zip\"";
-            response = Response.ok(zipFile, "application/zip").header("Content-Disposition",  contentDisposition).build();
         }
         return response;
     }
@@ -151,10 +153,12 @@ public class DownloadResource {
     @GET
     @Produces("application/json")
     public Response displayStagedItems() {
-        DownloadManager downloadManager = new DownloadManager();
-        List<Download> allStagedDownloads = downloadManager.getAllStagedDownloads();
-        Gson serializer = GsonUtil.getDefault();
-        String downloadJson = serializer.toJson(allStagedDownloads, ArrayList.class);
+        String downloadJson = null;
+        try (DownloadManager downloadManager = new DownloadManager()) {
+            List<Download> allStagedDownloads = downloadManager.getAllStagedDownloads();
+            Gson serializer = GsonUtil.getDefault();
+            downloadJson = serializer.toJson(allStagedDownloads, ArrayList.class);
+        }
         return Response.ok(downloadJson, MediaType.APPLICATION_JSON_TYPE).build();
     }
     
@@ -164,17 +168,21 @@ public class DownloadResource {
     public Response deleteStagedItem(@PathParam("itemId") String itemId, @Context HttpServletRequest request) {
         Response response = null;
         if (SessionResource.isValidSession(request)) {
-            DownloadManager downloadManager = new DownloadManager();
-            Download download = downloadManager.load(itemId);
-            boolean deleted = false;
-            try {
-                File stagingFolder = download.fetchZipFile().getParentFile();
-                deleted = FileUtils.deleteQuietly(stagingFolder);
-            } catch (URISyntaxException e) {
-                throw new RuntimeException(e);
+            try (DownloadManager downloadManager = new DownloadManager()) {
+                Download download = downloadManager.load(itemId);
+                boolean deleted = false;
+                try {
+                    if (download == null) {
+                        throw new NotFoundException();
+                    }
+                    File stagingFolder = download.fetchZipFile().getParentFile();
+                    deleted = FileUtils.deleteQuietly(stagingFolder);
+                } catch (URISyntaxException e) {
+                    throw new RuntimeException(e);
+                }
+                downloadManager.delete(download);
+                response = Response.ok("{\"deleted\":\"" + deleted + "\"}", MediaType.APPLICATION_JSON_TYPE).build();
             }
-            downloadManager.delete(download);
-            response = Response.ok("{\"deleted\":\"" + deleted + "\"}", MediaType.APPLICATION_JSON_TYPE).build();
         } else {
             throw new UnauthorizedException();
         }
